requirements.txt
fastapi
uvicorn
vision-parse[gemini]
python-multipart
img2pdf
Pillow